add exceptionController

+++ src/main/java/com/example/sklad/controller/GroupController.java
@@ -2,8 +2,12 @@ package com.example.sklad.controller;
 
 import com.example.sklad.dto.GroupShortDto;
 import com.example.sklad.entity.Group;
+import com.example.sklad.exceptions.NotFoundByIdException;
+import com.example.sklad.exceptions.NotUniqueNameException;
 import com.example.sklad.service.GroupService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,30 +21,33 @@ public class GroupController {
     GroupService groupService;
 
     @DeleteMapping("/delete/{id}")
-    public String delete(@PathVariable("id") UUID id) {
+    public ResponseEntity<?> delete(@PathVariable("id") UUID id) {
+        if (groupService.findById(id) == null) throw new NotFoundByIdException();
         groupService.deleteGroup(id);
-
-        return "delete";
+        return new ResponseEntity<>("Deleted successfully", HttpStatus.OK);
     }
 
-    // do recursion
     @GetMapping("/get/all")
-    public List<Group> getAll(){
-       return groupService.findAll();
+    public List<Group> getAll() {
+        return groupService.findAll();
     }
 
     @GetMapping("/get/{id}")
-    public Group getById(@PathVariable("id") UUID id){
-        return groupService.findById(id);
+    public Group getById(@PathVariable("id") UUID id) {
+        Group group = groupService.findById(id);
+        if (group == null) throw new NotFoundByIdException();
+        return group;
     }
 
     @PostMapping("/post/add")
-    public Group addNewGroup(@RequestBody GroupShortDto groupDto){
-        Group group = new Group();
-        group.setName(groupDto.getName());
-        group.setDescription(groupDto.getDescription());
-        return groupService.save(group);
+    public Group addNewGroup(@RequestBody GroupShortDto groupDto) {
+        boolean isNotUniqueName = groupService.findAll().stream()
+                .anyMatch(g -> g.getName().equals(groupDto.getName()));
+        if (isNotUniqueName) throw new NotUniqueNameException();
+        return groupService.save(Group.builder()
+                .name(groupDto.getName())
+                .description(groupDto.getDescription())
+                .build());
     }
-
 
 }
